Change return type of get_unix_time function from float to int

+++ cogs/utils/functions.py
@@ -414,7 +414,7 @@ def checkForDir(filepath):
             print (f"{filepath} created")
 
 def get_unix_time():
-    return float(time.time())
+    return int(time.time())
 
 
 def filter(message):
